# contracts/openapi_stub.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WAYMARK API",
    "description": "Explainable road risk, cell by cell, plus the action loop: region reports, an assistant, tracked measures with photo evidence and a verified-only adjusted risk estimate. Decision support only; human experts decide.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/cells": {
      "get": {
        "tags": [
          "cells"
        ],
        "summary": "List Cells",
        "operationId": "list_cells_api_cells_get",
        "parameters": [
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "name": "max_past_crashes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Past Crashes"
            }
          },
          {
            "name": "confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "High, Medium, Low (comma-separated allowed)",
              "title": "Confidence"
            },
            "description": "High, Medium, Low (comma-separated allowed)"
          },
          {
            "name": "emerging_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Emerging Only"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "south,west,north,east",
              "title": "Bbox"
            },
            "description": "south,west,north,east"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20000,
              "minimum": 1,
              "default": 10000,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cells/{cell_id}": {
      "get": {
        "tags": [
          "cells"
        ],
        "summary": "Cell Detail",
        "operationId": "cell_detail_api_cells__cell_id__get",
        "parameters": [
          {
            "name": "cell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cell Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/demo": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Demo",
        "description": "Demo-mode sign-in: a token for the chosen role. Switched off with WAYMARK_DEMO_MODE=0.",
        "operationId": "demo_api_auth_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CellDetail": {
        "properties": {
          "cell_id": {
            "type": "string",
            "title": "Cell Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "n_past_crashes": {
            "type": "integer",
            "title": "N Past Crashes"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "risk_vs_similar_history": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Vs Similar History"
          },
          "history_percentile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "History Percentile"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "confidence_explanation": {
            "type": "string",
            "title": "Confidence Explanation"
          },
          "emerging_risk": {
            "type": "boolean",
            "title": "Emerging Risk"
          },
          "top_overall": {
            "type": "boolean",
            "title": "Top Overall"
          },
          "recommend_audit": {
            "type": "boolean",
            "title": "Recommend Audit"
          },
          "audit_banner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audit Banner"
          },
          "baseline_prob": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Prob"
          },
          "top_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Factors"
          },
          "suggested_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Actions"
          },
          "shap": {
            "items": {
              "$ref": "#/components/schemas/ShapItem"
            },
            "type": "array",
            "title": "Shap"
          },
          "shap_available": {
            "type": "boolean",
            "title": "Shap Available"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioItem"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "google_maps_url": {
            "type": "string",
            "title": "Google Maps Url"
          },
          "adjusted_risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjusted Risk Score"
          },
          "has_measures": {
            "type": "boolean",
            "title": "Has Measures",
            "default": false
          },
          "measures": {
            "items": {
              "$ref": "#/components/schemas/CellMeasureBrief"
            },
            "type": "array",
            "title": "Measures",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "cell_id",
          "lat",
          "lng",
          "n_past_crashes",
          "risk_score",
          "risk_vs_similar_history",
          "history_percentile",
          "headline",
          "confidence",
          "confidence_explanation",
          "emerging_risk",
          "top_overall",
          "recommend_audit",
          "audit_banner",
          "baseline_prob",
          "top_factors",
          "suggested_actions",
          "shap",
          "shap_available",
          "scenarios",
          "google_maps_url"
        ],
        "title": "CellDetail"
      },
      "CellList": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CellLite"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "count",
          "limit",
          "offset",
          "items"
        ],
        "title": "CellList"
      },
      "CellLite": {
        "properties": {
          "cell_id": {
            "type": "string",
            "title": "Cell Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "n_past_crashes": {
            "type": "integer",
            "title": "N Past Crashes"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "risk_vs_similar_history": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Vs Similar History"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "emerging_risk": {
            "type": "boolean",
            "title": "Emerging Risk"
          },
          "top_overall": {
            "type": "boolean",
            "title": "Top Overall"
          },
          "adjusted_risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjusted Risk Score"
          },
          "has_measures": {
            "type": "boolean",
            "title": "Has Measures",
            "default": false
          },
          "measures": {
            "items": {
              "$ref": "#/components/schemas/CellMeasureBrief"
            },
            "type": "array",
            "title": "Measures",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "cell_id",
          "lat",
          "lng",
          "n_past_crashes",
          "risk_score",
          "risk_vs_similar_history",
          "confidence",
          "emerging_risk",
          "top_overall"
        ],
        "title": "CellLite"
      },
      "CellMeasureBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "region_kind": {
            "type": "string",
            "title": "Region Kind"
          },
          "region_key": {
            "type": "string",
            "title": "Region Key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "category",
          "region_kind",
          "region_key"
        ],
        "title": "CellMeasureBrief",
        "description": "Compact measure shown on map cells (additive field on /api/cells)."
      },
...
      "DemoIn": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "DemoIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "planner",
          "engineer",
          "community"
        ],
        "title": "Role"
      },
      "ScenarioItem": {
        "properties": {
          "night": {
            "type": "boolean",
            "title": "Night"
          },
          "rain": {
            "type": "boolean",
            "title": "Rain"
          },
          "low_vis": {
            "type": "boolean",
            "title": "Low Vis"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          }
        },
        "type": "object",
        "required": [
          "night",
          "rain",
          "low_vis",
          "probability"
        ],
        "title": "ScenarioItem"
      },
      "ShapItem": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "feature_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Value"
          },
          "shap_value": {
            "type": "number",
            "title": "Shap Value"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "label",
          "feature_value",
          "shap_value",
          "direction",
          "rank"
        ],
        "title": "ShapItem"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "TokenOut"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}